+++ EpicLoot/src/Adventure/feature/TreasureMaps.cs
@@ -108,6 +108,8 @@ namespace EpicLoot.Adventure.Feature
                 return false;
             }
 
+            PlayerExtensions_Adventure.PersistLocalPlayerAdventureData();
+
             Quaternion rotation = Quaternion.Euler(0f, UnityEngine.Random.Range(0f, 360f), 0f);
             GameObject gameObject = PrefabManager.Instance.GetPrefab(AdventureSpawnController.PrefabName);
             GameObject created_go = Object.Instantiate(gameObject, spawnPoint, rotation);